+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".App"
+        android:name="._unsorted.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -16,20 +16,20 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MediaActivity"
+            android:name="._unsorted.ui.MediaActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.SearchActivity"
+            android:name="._unsorted.ui.SearchActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustNothing"/>
         <activity
-            android:name=".ui.SettingsActivity"
+            android:name="._unsorted.ui.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.AudioplayerActivity"
+            android:name="._unsorted.ui.AudioplayerActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name="._unsorted.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
