+++ app/src/main/java/com/practicum/playlistmaker/_unsorted/App.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker._unsorted
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
