+++ app/src/main/java/com/practicum/playlistmaker/_unsorted/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.ui
+package com.practicum.playlistmaker._unsorted.ui
 
 import android.content.Intent
 import android.os.Bundle
