+++ app/src/main/java/com/practicum/playlistmaker/_unsorted_data/dto/TracksResponse.kt
@@ -1,6 +1,6 @@
-package com.practicum.playlistmaker.data.dto
+package com.practicum.playlistmaker._unsorted_data.dto
 
-import com.practicum.playlistmaker.domain.model.Track
+import com.practicum.playlistmaker._unsorted_domain.model.Track
 
 class TracksResponse(val searchType: String,
                      val expression: String,
