+++ app/src/main/java/com/practicum/playlistmaker/_unsorted_domain/repository/MyCallback.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker._unsorted_domain.repository
+
+interface MyCallback {
+    fun execute(message: String)
+}
