--- app/src/main/java/com/practicum/playlistmaker/domain/repository/MyCallback.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.domain.repository
-
-interface MyCallback {
-    fun execute(message: String)
-}
